requirements/base.txt
pytz==2019.1  # https://github.com/stub42/pytz
python-slugify==3.0.2  # https://github.com/un33k/python-slugify
Pillow==6.0.0  # https://github.com/python-pillow/Pillow
argon2-cffi==19.1.0  # https://github.com/hynek/argon2_cffi
redis==3.2.1  # https://github.com/antirez/redis
celery==4.3.0  # pyup: < 5.0  # https://github.com/celery/celery

# Django
# ------------------------------------------------------------------------------
django==2.2.3  # pyup: < 2.1  # https://www.djangoproject.com/
django-environ==0.4.5  # https://github.com/joke2k/django-environ
django-model-utils==3.1.2  # https://github.com/jazzband/django-model-utils
django-redis==4.10.0  # https://github.com/niwinz/django-redis

django-extensions==2.1.6  # https://github.com/django-extensions/django-extensions
django-mptt==0.9.1
easy-thumbnails==2.5
django_summernote==0.8.11.4

# Django REST Framework
djangorestframework==3.9.2  # https://github.com/encode/django-rest-framework
coreapi==2.3.3  # https://github.com/core-api/python-client
django-filter==2.0.0
